requirements.txt
fastapi
uvicorn[standard]
sqlmodel
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
psycopg2-binary
